# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/evaluate": {
      "post": {
        "description": "Resolves one flag for one user in one environment.\n\n**Deterministic:** the same `user_id` always receives the same value for a given flag configuration. Bucketing is a SHA-256 hash of the flag key and user id mapped to 0–99, so no state is stored per user and any instance computes the same answer.\n\n**Tenant scoping:** the tenant comes from the API key. `tenant_id` in the body is optional and only cross-checked.\n\nReturns `200` with the resolved value; `reason` explains why, and is never masked by caching.",
        "operationId": "EvaluationController_evaluate",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvaluateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Evaluated value.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvaluationResultDto"
                }
              }
            }
          },
          "400": {
            "description": "`flag_key` is missing, or `environment` is not one of development / staging / production.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid `x-api-key`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "A `tenant_id` was supplied in the body and does not match the authenticated API key.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "No active flag with that key for this tenant.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "Per-tenant rate limit exceeded. Limits are keyed on the API key so one tenant cannot consume another tenant's budget.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "Evaluate a single flag for a user",
        "tags": [
          "Evaluation"
        ]
      }
    },
    "/api/v1/evaluate/bulk": {
      "post": {
        "description": "Resolves all of the tenant's active (non-archived) flags for one user in a single call — the shape an SDK wants at client start-up.\n\n`flag_key` is ignored here. Archived flags are excluded.\n\nThis costs no more database round trips than a single evaluation: the tenant's flag definitions are fetched once from cache and every flag is then evaluated in memory.",
        "operationId": "EvaluationController_evaluateBulk",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvaluateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "All active flags, evaluated for this user.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkEvaluationResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "`environment` is not one of development / staging / production, or `user_id` is missing.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid `x-api-key`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "A `tenant_id` was supplied in the body and does not match the authenticated API key.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "Per-tenant rate limit exceeded. Limits are keyed on the API key so one tenant cannot consume another tenant's budget.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "Evaluate every active flag for a user",
        "tags": [
          "Evaluation"
        ]
      }
    },
    "/api/v1/tenants/{tenantId}/flags/{flagKey}/history": {
      "get": {
        "description": "Chronological change history for one flag, newest first.\n\nEvery create, update and archive is recorded automatically with who changed it, when, and the complete flag state before and after. Records are append-only — nothing in the service ever issues an UPDATE or DELETE against this table, and archiving a flag preserves its history rather than removing it.",
        "operationId": "AuditController_getHistory",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "path",
            "description": "Must match the tenant behind the API key.",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "flagKey",
            "required": true,
            "in": "path",
            "description": "Flag key.",
            "schema": {
              "example": "new-checkout-flow",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Audit entries, newest first. Empty array if the flag has no history or does not exist.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AuditLogDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid `x-api-key`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "The API key belongs to a different tenant.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "api-key": []
          }
        ],
        "summary": "Get a flag's change history",
        "tags": [
          "Audit"
        ]
      }
    },
...
  },
  "info": {
    "title": "Multi-Tenant Feature Flag Service",
...
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local (Docker Compose)"
    },
    {
      "url": "https://feature-flag-service-staging-dhqwxiey3q-uc.a.run.app",
      "description": "Staging (Cloud Run)"
    },
    {
      "url": "https://feature-flag-service-production-<hash>-uc.a.run.app",
      "description": "Production (Cloud Run)"
    }
  ],
  "components": {
    "securitySchemes": {
      "api-key": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "Tenant API key, returned once at registration. Identifies the tenant for every scoped operation."
      },
      "admin-key": {
        "type": "apiKey",
        "in": "header",
        "name": "x-admin-key",
        "description": "Platform admin key, required for tenant registration wherever ADMIN_API_KEY is configured. Read it with: gcloud secrets versions access latest --secret=feature-flag-service-<env>-admin-api-key"
      }
    },
    "schemas": {
...
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 400
          },
          "message": {
            "description": "Either a single message or, for validation failures, one entry per invalid field.",
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            ],
            "example": [
              "environment must be one of the following values: development, staging, production"
            ]
          },
          "error": {
            "type": "string",
            "example": "Bad Request"
          }
        },
        "required": [
          "statusCode",
          "message",
          "error"
        ]
      },
...
      "Environment": {
        "type": "string",
        "enum": [
          "development",
          "staging",
          "production"
        ],
        "description": "Environment this configuration applies to."
      },
...
      "EvaluateDto": {
        "type": "object",
        "properties": {
          "environment": {
            "description": "Environment to evaluate against.",
            "example": "production",
            "allOf": [
              {
                "$ref": "#/components/schemas/Environment"
              }
            ]
          },
          "user_id": {
            "type": "string",
            "description": "Stable identifier for the end user being evaluated. Drives deterministic bucketing — the same user always receives the same value for a given flag configuration. Also accepted as `userId`.",
            "example": "user-7a3f9b"
          },
          "tenant_id": {
            "type": "string",
            "format": "uuid",
            "description": "Optional. The tenant is always derived from the API key; if supplied this is only cross-checked against it and returns 403 on mismatch. It is never trusted as the source of tenant identity. Also accepted as `tenantId`.",
            "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
          },
          "context": {
            "type": "object",
            "description": "Arbitrary user attributes. Accepted and reserved for attribute-based targeting; not yet consulted during evaluation.",
            "additionalProperties": true,
            "example": {
              "country": "US",
              "plan": "enterprise"
            }
          },
          "flag_key": {
            "type": "string",
            "description": "Flag to evaluate. REQUIRED for POST /evaluate; ignored by POST /evaluate/bulk, which evaluates every active flag. Also accepted as `flagKey`.",
            "example": "new-checkout-flow"
          }
        },
        "required": [
          "environment",
          "user_id"
        ]
      },
      "EvaluationReason": {
        "type": "string",
        "enum": [
          "ENABLED",
          "FLAG_DISABLED",
          "NOT_IN_ROLLOUT",
          "VARIANT_MATCH"
        ],
        "description": "Why the flag resolved this way. Always the real cause — a cache hit does not mask it, because the cache stores flag definitions rather than results."
      },
      "EvaluationResultDto": {
        "type": "object",
        "properties": {
          "flagKey": {
            "type": "string",
            "example": "new-checkout-flow"
          },
          "value": {
            "description": "Resolved value for this user. Type follows the flag type.",
            "example": true,
            "oneOf": [
              {
                "type": "boolean"
              },
              {
                "type": "string"
              },
              {
                "type": "number"
              }
            ]
          },
          "reason": {
            "description": "Why the flag resolved this way. Always the real cause — a cache hit does not mask it, because the cache stores flag definitions rather than results.",
            "example": "ENABLED",
            "allOf": [
              {
                "$ref": "#/components/schemas/EvaluationReason"
              }
            ]
          }
        },
        "required": [
          "flagKey",
          "value",
          "reason"
        ]
      },
      "BulkEvaluationResponseDto": {
        "type": "object",
        "properties": {
          "flags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EvaluationResultDto"
            }
          },
          "count": {
            "type": "number",
            "description": "Number of flags evaluated.",
            "example": 5
          }
        },
        "required": [
          "flags",
          "count"
        ]
      },
      "AuditLogDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "tenantId": {
            "type": "string",
            "format": "uuid"
          },
          "flagId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "flagKey": {
            "type": "string",
            "example": "new-checkout-flow"
          },
          "action": {
            "type": "string",
            "description": "What happened.",
            "enum": [
              "CREATED",
              "UPDATED",
              "ARCHIVED"
            ],
            "example": "UPDATED"
          },
          "changedBy": {
            "type": "string",
            "description": "Identity that made the change.",
            "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
          },
          "previousValue": {
            "type": "object",
            "nullable": true,
            "description": "Full flag snapshot as it was BEFORE the change. Null for CREATED.",
            "additionalProperties": true
          },
          "newValue": {
            "type": "object",
            "nullable": true,
            "description": "Full flag snapshot as it was AFTER the change.",
            "additionalProperties": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "tenantId",
          "flagKey",
          "action",
          "changedBy",
          "previousValue",
          "newValue",
          "createdAt"
        ]
      },
...
    }
  }
}